pyproject.toml
[project]
name = "embeddings"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.29.5",
    "jupyter>=1.1.1",
    "plotly-express>=0.4.1",
    "plotly>=5.24.1",
    "pymongo>=4.10.1",
    "python-dotenv>=1.0.1",
    "ruff>=0.8.4",
    "scikit-learn>=1.6.0",
    "sentence-transformers>=3.3.1",
    "tqdm>=4.67.1",
    "pandas>=2.2.3",
    "requests>=2.32.3",
]
